refactor(examples/nextjs-app-optimistic-updates): apply post-review corrections
